Add new test: Checking to see if  still works if the repo is not a direct dependency of the root project, append the
